+++ earth2studio/perturbation/__init__.py
@@ -18,5 +18,6 @@ from .base import Perturbation
 from .brown import Brown  # noqa
 from .bv import BredVector  # noqa
 from .gaussian import Gaussian  # noqa
+from .lagged import LaggedEnsemble  # noqa
 from .spherical import SphericalGaussian  # noqa
 from .zero import Zero  # noqa