+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Perturbation method section in userguide
 - WeatherBench Climatology data source
 - Added `datasource_to_file` utility function
+- Add lagged ensemble perturbation method
 
 ### Changed